refactor(test): move fixtures to conftest.py
