+++ protoletariat/tests/test_rewrite.py
@@ -1,5 +1,4 @@
 import importlib
-import json
 import subprocess
 import sys
 from pathlib import Path
@@ -54,61 +53,6 @@ def check_proto_out(out: Path) -> None:
 def test_build_import_rewrite(proto: str, dep: str, expected: str) -> None:
     old, new = build_import_rewrite(proto, dep)
     assert new == expected
-
-
-@pytest.fixture
-def cli() -> CliRunner:
-    return CliRunner()
-
-
-@pytest.fixture
-def this_proto(tmp_path: Path) -> Path:
-    code = """
-syntax = "proto3";
-
-import "other.proto";
-import "baz/bizz_buzz.proto";
-
-package protoletariat.test;
-
-message Test {
-    Other test = 1;
-    protoletariat.test.baz.BuzzBuzz baz = 2;
-}
-"""
-    p = tmp_path.joinpath("this.proto")
-    p.write_text(code)
-    return p
-
-
-@pytest.fixture
-def other_proto(tmp_path: Path) -> Path:
-    code = """
-syntax = "proto3";
-
-package protoletariat.test;
-
-message Other {}
-"""
-    p = tmp_path.joinpath("other.proto")
-    p.write_text(code)
-    return p
-
-
-@pytest.fixture
-def baz_bizz_buzz_other_proto(tmp_path: Path) -> Path:
-    code = """
-syntax = "proto3";
-
-package protoletariat.test.baz;
-
-message BuzzBuzz {}
-"""
-    baz = tmp_path.joinpath("baz")
-    baz.mkdir()
-    p = baz.joinpath("bizz_buzz.proto")
-    p.write_text(code)
-    return p
 
 
 def test_cli(
@@ -200,90 +144,6 @@ def test_cli(
     assert str(tmp_path) not in sys.path
 
 
-@pytest.fixture
-def thing1_text() -> str:
-    return """
-// thing1.proto
-syntax = "proto3";
-
-import "thing2.proto";
-
-package things;
-
-message Thing1 {
-  Thing2 thing2 = 1;
-}
-"""
-
-
-@pytest.fixture
-def thing1(thing1_text: str, tmp_path: Path) -> Path:
-    p = tmp_path.joinpath("thing1.proto")
-    p.write_text(thing1_text)
-    return p
-
-
-@pytest.fixture
-def thing2_text() -> str:
-    return """
-// thing2.proto
-syntax = "proto3";
-
-package things;
-
-message Thing2 {
-  string data = 1;
-}
-"""
-
-
-@pytest.fixture
-def thing2(thing2_text: str, tmp_path: Path) -> Path:
-    code = """
-// thing2.proto
-syntax = "proto3";
-
-package things;
-
-message Thing2 {
-  string data = 1;
-}
-"""
-    p = tmp_path.joinpath("thing2.proto")
-    p.write_text(code)
-    return p
-
-
-@pytest.fixture
-def thing1_nested_text() -> str:
-    return """
-// thing1.proto
-syntax = "proto3";
-
-import "a/b/c/thing2.proto";
-
-package thing1.a;
-
-message Thing1 {
-  thing2.a.b.c.Thing2 thing2 = 1;
-}
-"""
-
-
-@pytest.fixture
-def thing2_nested_text() -> str:
-    return """
-// thing2.proto
-syntax = "proto3";
-
-package thing2.a.b.c;
-
-message Thing2 {
-  string data = 1;
-}
-"""
-
-
 def test_example_protoc(
     cli: CliRunner,
     tmp_path: Path,
@@ -336,36 +196,11 @@ def test_example_buf(
     tmp_path: Path,
     thing1: Path,
     thing2: Path,
+    buf_yaml: Path,
+    buf_gen_yaml: Path,
+    out: Path,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    out = tmp_path.joinpath("out")
-    out.mkdir()
-
-    monkeypatch.chdir(tmp_path)
-    with tmp_path.joinpath("buf.yaml").open(mode="w") as f:
-        json.dump(
-            {
-                "version": "v1",
-                "lint": {"use": ["DEFAULT"]},
-                "breaking": {"use": ["FILE"]},
-            },
-            f,
-        )
-
-    with tmp_path.joinpath("buf.gen.yaml").open(mode="w") as f:
-        json.dump(
-            {
-                "version": "v1",
-                "plugins": [
-                    {
-                        "name": "python",
-                        "out": str(out),
-                    }
-                ],
-            },
-            f,
-        )
-
     subprocess.check_call(
         [
             "protoc",
@@ -381,6 +216,8 @@ def test_example_buf(
     )
 
     check_proto_out(out)
+
+    monkeypatch.chdir(tmp_path)
 
     result = cli.invoke(
         main,
@@ -404,13 +241,13 @@ def test_example_buf(
 def test_nested_buf(
     cli: CliRunner,
     tmp_path: Path,
+    buf_yaml: Path,
+    buf_gen_yaml_nested: Path,
+    out_nested: Path,
     thing1_nested_text: str,
     thing2_nested_text: str,
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
-    out = tmp_path.joinpath("out_nested")
-    out.mkdir()
-
     c = tmp_path / "a" / "b" / "c"
     c.mkdir(parents=True, exist_ok=True)
 
@@ -423,31 +260,6 @@ def test_nested_buf(
     thing1 = d / "thing1.proto"
     thing1.write_text(thing1_nested_text)
 
-    monkeypatch.chdir(tmp_path)
-    with tmp_path.joinpath("buf.yaml").open(mode="w") as f:
-        json.dump(
-            {
-                "version": "v1",
-                "lint": {"use": ["DEFAULT"]},
-                "breaking": {"use": ["FILE"]},
-            },
-            f,
-        )
-
-    with tmp_path.joinpath("buf.gen.yaml").open(mode="w") as f:
-        json.dump(
-            {
-                "version": "v1",
-                "plugins": [
-                    {
-                        "name": "python",
-                        "out": str(out),
-                    },
-                ],
-            },
-            f,
-        )
-
     subprocess.check_call(
         [
             "protoc",
@@ -456,17 +268,19 @@ def test_nested_buf(
             "--proto_path",
             str(tmp_path),
             "--python_out",
-            str(out),
+            str(out_nested),
         ],
     )
 
-    check_proto_out(out)
+    check_proto_out(out_nested)
+
+    monkeypatch.chdir(tmp_path)
 
     result = cli.invoke(
         main,
         [
             "-o",
-            str(out),
+            str(out_nested),
             "--in-place",
             "--create-package",
             "buf",
@@ -484,82 +298,11 @@ def test_nested_buf(
         importlib.import_module("out_nested.d.thing1_pb2")
 
     # check that we created packages in the correct locations
-    assert out.joinpath("__init__.py").exists()
-    assert out.joinpath("a", "__init__.py").exists()
-    assert out.joinpath("a", "b", "__init__.py").exists()
-    assert out.joinpath("a", "b", "c", "__init__.py").exists()
-    assert out.joinpath("d", "__init__.py").exists()
-
-
-@pytest.fixture
-def thing_service_text() -> str:
-    return """
-// thing_service.proto
-syntax = "proto3";
-
-import "thing1.proto";
-import "thing2.proto";
-
-package things;
-
-service ThingService {
-    rpc GetThing1(Thing1) returns (Thing1) {}
-    rpc GetThing2(Thing2) returns (Thing2) {}
-}
-"""
-
-
-@pytest.fixture
-def thing_service(thing_service_text: str, tmp_path: Path) -> Path:
-    p = tmp_path.joinpath("thing_service.proto")
-    p.write_text(thing_service_text)
-    return p
-
-
-@pytest.fixture
-def buf_yaml(tmp_path: Path) -> Path:
-    p = tmp_path.joinpath("buf.yaml")
-    p.write_text(
-        json.dumps(
-            {
-                "version": "v1",
-                "lint": {"use": ["DEFAULT"]},
-                "breaking": {"use": ["FILE"]},
-            },
-        )
-    )
-    return p
-
-
-@pytest.fixture
-def out_grpc(tmp_path: Path) -> Path:
-    out = tmp_path / "out_grpc"
-    out.mkdir()
-    return out
-
-
-@pytest.fixture
-def buf_gen_yaml_grpc(tmp_path: Path, out_grpc: Path) -> Path:
-    p = tmp_path.joinpath("buf.gen.yaml")
-    p.write_text(
-        json.dumps(
-            {
-                "version": "v1",
-                "plugins": [
-                    {
-                        "name": "python",
-                        "out": str(out_grpc),
-                    },
-                    {
-                        "name": "grpc_python",
-                        "out": str(out_grpc),
-                        "path": "grpc_python_plugin",
-                    },
-                ],
-            },
-        )
-    )
-    return p
+    assert out_nested.joinpath("__init__.py").exists()
+    assert out_nested.joinpath("a", "__init__.py").exists()
+    assert out_nested.joinpath("a", "b", "__init__.py").exists()
+    assert out_nested.joinpath("a", "b", "c", "__init__.py").exists()
+    assert out_nested.joinpath("d", "__init__.py").exists()
 
 
 def test_grpc_buf(
